Navigation handling from song > note
